pyproject.toml
[project]
name = "voice-interview-sequential-google"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.9.0",
    "python-dotenv>=1.1.1",
]
